[feat][Angel] created a separate component file for Header under Profile and added a few functions in case page.tsx
