+++ src/app/dashboard/case/page.tsx
@@ -1,42 +1,14 @@
 "use client";
 import * as React from "react";
 import { IoMdCheckmark } from "react-icons/io";
-import { CiBellOn, CiSearch } from "react-icons/ci";
-import BaseFormSelect from "@/components/Global/BaseFormSelect";
-import BaseFormInput from "@/components/Global/BaseFormInput";
-import { Case, Category } from "@/interface/CaseTypes";
+import { cases, Category } from "@/interface/CaseTypes";
 import CaseCard from "@/app/dashboard/case/CaseCard";
-import { useEffect, useRef } from "react";
-import { useRouter } from "next/navigation";
+import BaseFormSelect from "@/components/Global/BaseFormSelect";
+import Header from "@/components//Profile/Header";
 
 const sortingOptions = [
   { label: "Latest first", value: "latest" },
   { label: "Oldest first", value: "oldest" },
-];
-
-const cases: Case[] = [
-  {
-    id: "1",
-    title: "Need Help in Leaving my Husband",
-    category: { name: "Divorce Cases" },
-    status: { isOpen: true },
-    avatar:
-      "https://cdn.builder.io/api/v1/image/assets/TEMP/05f94ecc6aa9da0cde7bbf9b88c7a8f8bff72e91?placeholderIfAbsent=true&apiKey=b97adb845ad745fdabf283f95e3c166e",
-    lastUpdate: { user: "@chimichangas", time: "30 minutes ago" },
-    description:
-      "Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum",
-  },
-  {
-    id: "2",
-    title: "Tabang unta ma akoa among yuta",
-    category: { name: "Land Ownership" },
-    status: { isOpen: false },
-    avatar:
-      "https://cdn.builder.io/api/v1/image/assets/TEMP/05f94ecc6aa9da0cde7bbf9b88c7a8f8bff72e91?placeholderIfAbsent=true&apiKey=b97adb845ad745fdabf283f95e3c166e",
-    lastUpdate: { user: "@chimichangas", time: "1 hour ago" },
-    description:
-      "Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum",
-  },
 ];
 
 const categories: Category[] = [
@@ -48,91 +20,12 @@ const categories: Category[] = [
   { id: "human", name: "Human Rights", color: "bg-pink-600" },
 ];
 
-const SearchBar: React.FC<{ searchQuery: string; setSearchQuery: React.Dispatch<React.SetStateAction<string>> }> = ({
-  searchQuery,
-  setSearchQuery,
-}) => {
-  return (
-    <div className="relative flex items-center border border-black rounded-[40px] mt-7 px-5 py-2.5 w-full h-10 
-      max-w-2lg 
-      sm:max-w-sm
-      md:max-w-md
-      lg:max-w-lg
-      xl:max-w-xl"
-    >
-      <BaseFormInput
-        label=""
-        name="search"
-        type="text"
-        placeholder="Search"
-        value={searchQuery}
-        onChange={(e) => setSearchQuery(e.target.value)}
-        className="px-4 py-2 mb-5 focus:outline-none focus:ring-0 focus:border-transparent w-full"
-      />
-      <CiSearch className="absolute right-4 text-blue-500 text-xl" />
-    </div>
-  );
-};
-
-const NotificationIcon: React.FC<{ notificationCount: number; isOpen: boolean; setIsOpen: React.Dispatch<React.SetStateAction<boolean>> }> = ({ notificationCount, isOpen, setIsOpen }) => {
-  const notificationRef = useRef<HTMLDivElement>(null);
-
-  const togglePreview = () => {
-    setIsOpen((prev) => !prev);
-  };
-
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (notificationRef.current && !notificationRef.current.contains(event.target as Node)) {
-        setIsOpen(false);
-      }
-    };
-
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-    };
-  }, []);
-
-  return (
-    <div className="relative inline-block select-none" ref={notificationRef}>
-      <div onClick={togglePreview} className="relative cursor-pointer">
-        <CiBellOn size={28} className="text-black mt-8" />
-        {notificationCount > 0 && (
-          <>
-            <span className="z-10 absolute w-2 h-2 top-1 right-1 bg-[#B32828] rounded-full"></span>
-          </>
-        )}
-      </div>
-
-      {isOpen && (
-        <div className="absolute right-0 mt-2 w-64 bg-white shadow-md rounded-lg p-3 border border-gray-200 z-[9999]">
-          <p className="text-sm font-semibold">Notifications</p>
-          <ul className="mt-2">
-            {cases.length > 0 ? (
-              cases.map((c) => (
-                <li key={c.id} className="flex items-center space-x-3 text-xs text-gray-700 py-2 border-b last:border-b-0">
-                  <img src={c.avatar} alt="avatar" className="w-8 h-8 rounded-full object-cover" />
-                  <div>
-                    <p className="font-semibold">{c.title}</p>
-                    <p className="text-gray-500 text-[10px]">{c.lastUpdate.user} • {c.lastUpdate.time}</p>
-                  </div>
-                </li>
-              ))
-            ) : (
-              <li className="text-xs text-gray-500 py-2 text-center">No new notifications</li>
-            )}
-          </ul>
-        </div>
-      )}
-    </div>
-  );
-};
-
 const Sidebar: React.FC<{
   selectedCaseType: string;
   setSelectedCaseType: (type: string) => void;
-}> = ({ selectedCaseType, setSelectedCaseType }) => (
+  selectedCategory: string | null;
+  setSelectedCategory: (category: string | null) => void;
+}> = ({ selectedCaseType, setSelectedCaseType, selectedCategory, setSelectedCategory }) => (
   <aside
     className="ml-5 w-[23%] max-md:ml-0 max-md:w-full"
     role="complementary"
@@ -186,10 +79,17 @@ const Sidebar: React.FC<{
         <hr className="mt-3 border-black border-opacity-30" />
         <ul className="mt-5" role="list">
           {categories.map((category) => (
-            <li key={category.id} className="flex gap-3 mt-6 ml-3.5 hover:underline">
+            <li 
+              key={category.id} 
+              className={`flex gap-3 mt-6 ml-3.5 hover:underline cursor-pointer ${
+                selectedCategory === category.name ? "text-[#0838E5] font-bold" : "text-black"
+              }`}
+              onClick={() => setSelectedCategory(selectedCategory === category.name ? null : category.name)}
+            >
               <span className={`flex self-center shrink-0 w-2 h-2 ${category.color} rounded-full`} aria-hidden="true" />
               <span>{category.name}</span>
-            </li>
+              {selectedCategory === category.name && <IoMdCheckmark className="ml-auto text-blue-600 text-xl" />}
+            </li>          
           ))}
         </ul>
     </nav>
@@ -197,13 +97,13 @@ const Sidebar: React.FC<{
 );
 
 const InputDesign: React.FC = () => {
-  const router = useRouter();
-  const [sortOrder, setSortOrder] = React.useState("latest");
   const [searchQuery, setSearchQuery] = React.useState("");
-  const [isOpen, setIsOpen] = React.useState(false);
+  const [sortOrder, setSortOrder] = React.useState("latest");
   const [selectedCaseType, setSelectedCaseType] = React.useState("all");
-  // const [username, setUsername] = useState<string | null>(null); //needed for the real-time username fetching commented out only kay temp ra ang data
   const openCaseCount = cases.filter((c) => c.status.isOpen).length;
+  const [selectedCategory, setSelectedCategory] = React.useState<string | null>(null);
+
+  // const [username, setUsername] = useState<string | null>(null); //needed for the real-time username fetching commented out only kay temp ra ang data
 
   // useEffect(() => { 
   //   const fetchUser = async () => {
@@ -221,58 +121,35 @@ const InputDesign: React.FC = () => {
 
   const filteredCases = cases.filter((caseItem) => {
     const query = searchQuery.toLowerCase();
-    
-    return (
+  
+    const matchesCaseType =
+      selectedCaseType === "all"
+        ? true
+        : selectedCaseType === "open"
+        ? caseItem.status.isOpen
+        : !caseItem.status.isOpen;
+  
+    const matchesCategory =
+      selectedCategory === null || caseItem.category.name === selectedCategory;
+  
+    const matchesSearch =
       caseItem.title.toLowerCase().includes(query) ||
-      caseItem.category.name.toLowerCase().includes(query) ||
-      (query === "open" && caseItem.status.isOpen) ||
-      ((query === "closed" || query === "close") && !caseItem.status.isOpen)
-    );
+      caseItem.category.name.toLowerCase().includes(query);
+  
+    return matchesCaseType && matchesCategory && matchesSearch;
   });  
 
   const sortedCases = [...filteredCases].sort((a, b) => {
-    if (sortOrder === "latest") {
-      return b.lastUpdate.time.localeCompare(a.lastUpdate.time); 
-    } else {
-      return a.lastUpdate.time.localeCompare(b.lastUpdate.time); 
-    }
+      if (sortOrder === "latest") {
+        return b.lastUpdate.time.localeCompare(a.lastUpdate.time); 
+      } else {
+        return a.lastUpdate.time.localeCompare(b.lastUpdate.time); 
+      }
   });
 
   return (
     <main className="flex flex-col text-black w-full max:w-100vw font-[Poppins]" role="main">
-      <header className="p-2.5 w-full bg-white rounded-[30px_30px_0px_0px] shadow-[0px_4px_4px_rgba(0,0,0,0.25)]">
-        <div className="flex max-md:flex-col">
-          <div className="w-[77%] max-md:w-full">
-            <div className="flex flex-wrap gap-6 text-xs">
-              <img
-                src="https://cdn.builder.io/api/v1/image/assets/TEMP/5676b5cbc1d09b6b170298efcc84d833fd17cbb2?placeholderIfAbsent=true&apiKey=b97adb845ad745fdabf283f95e3c166e"
-                alt="Logo"
-                className="w-[139px] object-contain cursor-pointer"
-                onClick={() => router.push("/dashboard/home")}
-              />
-              <SearchBar searchQuery={searchQuery} setSearchQuery={setSearchQuery} />
-            </div>
-          </div>
-          <div className="flex gap-5">
-            <div className={`relative transition-transform duration-100 ${isOpen ? '' : 'hover:scale-90'}`}>
-              <NotificationIcon notificationCount={openCaseCount} isOpen={isOpen} setIsOpen={setIsOpen} />
-            </div>
-            <div
-              className="z-0 flex position-fixed gap-4 ml-5 items-start text-xs font-medium cursor-pointer transition-transform duration-200 hover:scale-105"
-              onClick={() => router.push("/dashboard/account")}
-              >
-              <img
-                src="https://cdn.builder.io/api/v1/image/assets/TEMP/e5990c8fddf80298658964bf5d965e28f274456d?placeholderIfAbsent=true&apiKey=b97adb845ad745fdabf283f95e3c166e"
-                alt="Profile"
-                className="z-10 w-[60px] rounded-full mt-3"
-                />
-              {/* <span className="my-auto">{username || "Guest"}</span> */}
-              <span className="my-auto">rexhermoso</span> {/*this is temporary */}
-            </div>
-          </div>
-        </div>
-      </header>
-
+      <Header searchQuery={searchQuery} setSearchQuery={setSearchQuery} openCaseCount={openCaseCount} />
       <section className="self-center mt-10 pb-10 w-full max-w-[976px] h-[calc(100vh-40px)] max-h-[65vh] flex flex-col" aria-label="Case listings">
         <div className="flex gap-5 max-md:flex-col pb-5 overflow-y-auto overflow-x-hidden">
           <div className="w-[77%] max-md:w-full">
@@ -292,7 +169,12 @@ const InputDesign: React.FC = () => {
               )}
             </div>
           </div>
-          <Sidebar selectedCaseType={selectedCaseType} setSelectedCaseType={setSelectedCaseType} />
+          <Sidebar 
+            selectedCaseType={selectedCaseType} 
+            setSelectedCaseType={setSelectedCaseType} 
+            selectedCategory={selectedCategory}
+            setSelectedCategory={setSelectedCategory}
+          />
         </div>
       </section>
     </main>
